folder amd import refactor
