+++ client/src/components/ListItem/ListItem.js
@@ -1,8 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import styles from "./ListItem.module.css";
-import SwapButton from "../UI/SwapButton/SwapButton";
-import DeleteButton from "../UI/DeleteButton/DeleteButton";
+import { SwapButton, DeleteButton} from "../../components"
 
 const listItem = ({ text, buttonText, id }) => (
   <div className={styles.listItem}>
