+++ client/src/components/index.js
@@ -1,0 +1,5 @@
+export { default as ListItem } from "./ListItem/ListItem";
+export { default as ShoppingInput } from "./ShoppingInput/ShoppingInput";
+export { default as ShoppingList } from "./ShoppingList/ShoppingList";
+export { default as DeleteButton } from "./UI/DeleteButton/DeleteButton";
+export { default as SwapButton } from "./UI/SwapButton/SwapButton";
