+++ client/src/containers/index.js
@@ -1,0 +1,2 @@
+export { default as Layout } from "./Layout/Layout";
+export { default as ListsBody } from "./ListsBody/ListsBody";
