+++ client/src/util/index.js
@@ -1,0 +1,2 @@
+export { default as alphabetize } from "./Alphabetize";
+export { default as inputValidation } from "./InputValidation";
